requirements.txt
Django==2.2.2
Pillow==6.0.0
python-telegram-bot==11.1.0
django-uuslug==1.1.8
django-crispy-forms==1.7.2